CompareLocaleのリスト表示をFragment化
